refactor(mc/tui): group inline image code under components/images
Move image-manager.ts into components/images/ and extract
InlineImageComponent (previously inlined at the top of
tool-execution-enhanced.ts) into its own file alongside it. They form
a tight pair — the component implements ImageOwner and registers with
the manager — so colocating them keeps the lifecycle in one place and
shrinks tool-execution-enhanced.ts.

No behavior change.
