+++ mastracode/src/tui/components/images/image-manager.ts
