+++ mastracode/src/tui/components/images/index.ts
@@ -1,0 +1,2 @@
+export { imageManager, type ImageOwner } from './image-manager.js';
+export { InlineImageComponent, type InlineImageArgs } from './inline-image.js';
